Ensure timestamp for CloudWatch is the time of the log entry
Currently, the `timestamp` value that's used on the `log_events` objects is the
time at which the event object is created and sent to CloudWatch.

This does not always accurately reflect the `datetime` value of the entry
actually being logged in the logger. In the case of latency between a message
being pushed onto the queue (possibly a large queue), and it being popped off,
the times will be different.

The times in the actual log message will be accurate (see below), as they are
derived from `datetime`, but the epoch time sent to CloudWatch will differ.

Ensuring the time used for the `timestamp` value sent to CloudWatch is the time
when the entry was logged, will also help if in the future, batch posting
multiple entries is supported.

Below is an example illustrating how there can be a disparity between times
currently, and how this PR helps to resolve that:

**Before:**
```
[24/Jun/2019:14:28:59 -0400] "POST /log HTTP/1.1" 202 21 0.0044

2019-06-24 14:28:59 -0400 pid=1, thread=1, severity=INFO, {}

1561400960123 (2019-06-24 14:29:20.123)
```

**After:**
```
[24/Jun/2019:14:22:30 -0400] "POST /log HTTP/1.1" 202 21 0.0032

2019-06-24 14:22:30 -0400 pid=2, thread=2, severity=INFO, {}

1561400550831 (2019-06-24 14:22:30.831)
```

- determine epoch time from `datetime`
- during formatting, normalize an object that we can store this more accurate
  timestamp on
- reference the new `message_object[:epoch_time]` to set the `timestamp` on the
  CloudWatch entries
- reference the new `message_object[:message]` to pass through the message,
  unchanged