Ensure backwards compatibility when default log formatter is not used
In order for changes in #12 to not cause breakage, it assumes that the default
formatter is used, or that a custom formatter would have `epoch_time` and `message`
available on the `Hash` object.

This was surfaced by: https://github.com/zshannon/cloudwatchlogger/issues/16

These changes leave the default formatter changes in place, but ensure that the
`timestamp` and `message` values on the log events revert to the original values
prior to #12 if the message on the queue is not a `Hash` and does not contain
`epoch_time` and `message` values. This will allow backwards compatibility and
also allow for custom formatters to still pass these values in if they wish to
benefit from the timestamp disparity.